novos comandos

+++ commands/ban.js
@@ -36,9 +36,9 @@ module.exports = {
       }
       return;
     }
-    function verificacoes(members) {
+    function verificacoes(member) {
       let podeIr = false
-      if (members.has(message.guild.ownerID)) {
+      if (member.user.id === message.guild.ownerID) {
         if (podeEnviarMsg) {
           message.channel.send(`<:slashred:747879954305253468> ${message.author}, ele é o dono do servidor, não posso fazer isso!`)
         } else if (podeAddReactions) {
@@ -46,7 +46,7 @@ module.exports = {
         }
         return podeIr = false;
       }
-      if (members.map(user => user.roles.highest.position >= botMembro.roles.highest.position).indexOf(true) !== -1) {
+      if (member.roles.highest.position >= botMembro.roles.highest.position) {
         if (podeEnviarMsg) {
           message.channel.send(`<:slashred:747879954305253468> ${message.author}, eu não posso banir esse membro, ele tem um cargo maior que o meu!`)
         } else if (podeAddReactions) {
@@ -54,7 +54,7 @@ module.exports = {
         }
         return podeIr = false;
       }
-      if (members.map(user => user.roles.highest.position >= message.member.roles.highest.position).indexOf(true) !== -1 && message.author.id !== message.guild.ownerID) {
+      if (member.roles.highest.position >= message.member.roles.highest.position && message.author.id !== message.guild.ownerID) {
         if (podeEnviarMsg) {
           message.channel.send(`<:slashred:747879954305253468> ${message.author}, eu não posso banir esse membro, ele tem um cargo maior que o seu!`)
         } else if (podeAddReactions) {
@@ -62,7 +62,7 @@ module.exports = {
         }
         return podeIr = false;
       }
-      if (members.has(botMembro)) {
+      if (member.user.id === client.user.id) {
         if (podeEnviarMsg) {
           message.channel.send(`<:alertcircleamarelo:747879938207514645> ${message.author}, eu não posso me banir do servidor, faça isso manualmente ou peça ajuda a outro bot!`);
         } else if (podeAddReactions) {
@@ -70,7 +70,7 @@ module.exports = {
         }
         return podeIr = false;
       }
-      if (members.has(message.member)) {
+      if (member === message.member) {
         if (podeEnviarMsg) {
           message.channel.send(`<:alertcircleamarelo:747879938207514645> ${message.author}, você não pode se banir do servidor, isso é apenas questão de segurança!`);
         } else if (podeAddReactions) {
@@ -96,25 +96,25 @@ module.exports = {
         return;
       } else {
         for(let i = 0; i < usernamesDigitados.length; i++) {
-          const usernameMembers = await message.guild.members.cache.filter(member => member.user.username.toLowerCase() === usernamesDigitados[i])
-          const nicknameMembers = await message.guild.members.cache.filter(member => (member.nickname === null || member.nickname === undefined) ? member.nickname : member.nickname.toLowerCase() === usernamesDigitados[i])
-          if(usernameMembers.size !== 0) {
-            if(!verificacoes(usernameMembers))return;
-            usernameMembers.map(member => member.ban({reason: motivo, days: (Number(daysMsgDelete) >= 0 && Number(daysMsgDelete) <= 7) ? Number(daysMsgDelete) : 7}))
+          const usernameMember = await message.guild.members.cache.find(member => member.user.username.toLowerCase() === usernamesDigitados[i])
+          const nicknameMember = await message.guild.members.cache.find(member => (member.nickname === null || member.nickname === undefined) ? member.nickname : member.nickname.toLowerCase() === usernamesDigitados[i])
+          if(usernameMember) {
+            if(!verificacoes(usernameMember))return;
+            usernameMember.ban({reason: motivo, days: (Number(daysMsgDelete) >= 0 && Number(daysMsgDelete) <= 7) ? Number(daysMsgDelete) : 7})
             if (podeManageMessages && i === usernamesDigitados.length - 1) {
               message.delete();
-            } else if(podeEnviarMsg) {  
-              message.channel.send(`**${usernameMembers.map(member => member.user.username)[0]}** foi banido com sucesso! <:circlecheckverde:747879943224033481>`)
+            } else if(podeEnviarMsg) { 
+              message.channel.send(`<:circlecheckverde:747879943224033481> **${usernameMember.map(member => member.user.username)[0]}** foi banido com sucesso!`)
             } else if(podeAddReactions) {
               message.react('circlecheckverde:747879943224033481')
             }
-          } else if(nicknameMembers.size !== 0) {
-            if(!verificacoes(nicknameMembers))return;
-            nicknameMembers.map(member => member.ban({reason: motivo, days: (Number(daysMsgDelete) >= 0 && Number(daysMsgDelete) <= 7) ? Number(daysMsgDelete) : 7}))
+          } else if(nicknameMember) {
+            if(!verificacoes(nicknameMember))return;
+            nicknameMember.ban({reason: motivo, days: (Number(daysMsgDelete) >= 0 && Number(daysMsgDelete) <= 7) ? Number(daysMsgDelete) : 7})
             if (podeManageMessages && i === usernamesDigitados.length - 1) {
               message.delete();
             } else if(podeEnviarMsg) {
-              message.channel.send(`**${nicknameMembers.map(member => member.nickname)[0]}** foi banido com sucesso! <:circlecheckverde:747879943224033481>`)
+              message.channel.send(`<:circlecheckverde:747879943224033481> **${nicknameMember.map(member => member.nickname)[0]}** foi banido com sucesso!`)
             } else if(podeAddReactions) {
               message.react('circlecheckverde:747879943224033481')
             }
@@ -129,18 +129,16 @@ module.exports = {
         return;
       }
     }
-    if(!verificacoes(mencoes.members))return;
-    mencoes.members.map(user => user.ban({reason: motivo, days: (Number(daysMsgDelete) >= 0 && Number(daysMsgDelete) <= 7) ? Number(daysMsgDelete) : 7}))
-    if (podeManageMessages) {
-      message.delete();
-    } else if(podeEnviarMsg) {
-      if(mencoes.members.size > 1) {
-        message.reply(`**${mencoes.members.map(member => member.user.username).join(', ')}** foram banidos com sucesso! <:circlecheckverde:747879943224033481>`)
-      } else {
-        message.reply(`**${mencoes.members.map(member => member.user.username)[0]}** foi expulso com sucesso! <:circlecheckverde:747879943224033481>`)
+    for(let i = 0; i < mencoes.members.size; i++) {
+      if(!verificacoes(mencoes.members.map(x => x)[i]))return;
+      mencoes.members.map(x => x)[i].ban({reason: motivo, days: (Number(daysMsgDelete) >= 0 && Number(daysMsgDelete) <= 7) ? Number(daysMsgDelete) : 7})
+      if (podeManageMessages && i === mencoes.members.size - 1) {
+        message.delete();
+      } else if(podeEnviarMsg) {
+          message.reply(`<:circlecheckverde:747879943224033481> **${mencoes.members.map(x => x)[i].user.username}** foi expulso com sucesso!`)
+      } else if(podeAddReactions) {
+        message.react('circlecheckverde:747879943224033481')
       }
-    } else if(podeAddReactions) {
-      message.react('circlecheckverde:747879943224033481')
     }
   }
 }
